pyproject.toml
[tool.isort]
skip = ["env", "venv"]
profile = "black"
[tool.poetry]
name = "env_selenium"
version = "0.1.0"
description = "Selenium Python example project with pytest and Allure report"
authors = ["Margo"]
readme = "README.md"
packages = [{include = "env_selenium"}]

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "5.12.0"
pre-commit = "^3.4.0"

[tool.pytest.ini_options]
addopts = ["--alluredir=allure-results", "--color=yes", "-s", "-p no:logging"]
base_url = 'https://sts-q.in.top/'
operator_url = 'https://sts-q.in.top/Account/Registration?profileType=1'

markers = [
    "security: marks security tests",
    "devRun: marks tests that run before merge to the main branch",
]
testpaths = [
  "tests",
]

[tool.poetry.dependencies]
python = "^3.10"
allure-pytest = "2.13.2"
allure-python-commons = "2.13.2"
appdirs = "1.4.4"
assertpy = "1.1"
atomicwrites = "1.4.1"
attrs = "23.1.0"
certifi = "2023.7.22"
chardet = "5.2.0"
colorama = "0.4.6"
configparser = "6.0.0"
crayons = "0.4.0"
distlib = "0.3.7"
filelock = "3.12.4"
idna = "3.4"
iniconfig = "2.0.0"
packaging = "23.1"
pluggy = "1.2.0"
py = "1.11.0"
pyparsing = "3.1.1"
pytest = "7.4.2"
pytest-ordering = "0.6"
requests = "2.31.0"
selenium = "4.12.0"
six = "1.16.0"
toml = "0.10.2"
urllib3 = "2.0.4"
virtualenv = "20.24.5"
virtualenv-clone = "0.5.7"
xlrd = "2.0.1"
pytest-base-url = "^2.0.0"
python-dotenv = "^1.0.0"
mailinator-python-client-2 = "^0.0.4"
visual-regression-tracker = "^4.9.0"
pytest-check = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
